requirements.txt
requests
numpy
pandas
scipy
matplotlib
scikit-learn
flask
django
pytest
beautifulsoup4
pillow
sqlalchemy
lxml
jupyter
notebook
tqdm
seaborn
pyyaml
pytest-cov
ipython
sphinx
fastapi
uvicorn
streamlit
google-generativeai
python-dotenv
pydantic
requests-oauthlib
httpx
aiohttp
asyncio
websockets
pytest-asyncio
pytest-httpx
pytest-django
pytest-flask
pytest-cov
pytest-mock
pytest-xdist
streamlit
openai
python-dotenv
google-generativeai
